timetable_formatter/requirements.txt
pandas~=2.2.0rc0
tabula-py~=2.9.0
camelot-py~=0.11.0
opencv-python-headless~=4.9.0.80
python-magic~=0.4.27
python3-ghostscript~=0.5.0